Bugfix: axvspan in matplotlib<3.6 behavior change
axvspan.get_xy() used to return a numpy array of shape (5, 2) containing
the coordinates of the five vertices that make up the rectangular span
(the fifth vertex is the same as the first to close the shape). Starting
from matplotlib version 3.6, instead of returning all the vertices, it
now returns a simple Python tuple (x, y) representing the coordinates of
the bottom-left corner.